Move Theme to ui module
