+++ features/shopping_lists/src/main/kotlin/gq/kirmanak/mealient/shopping_lists/ui/composables/CenteredText.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import gq.kirmanak.mealient.AppTheme
+import gq.kirmanak.mealient.ui.AppTheme
 
 @Composable
 fun CenteredText(
