+++ features/shopping_lists/src/main/kotlin/gq/kirmanak/mealient/shopping_lists/ui/composables/EmptyListError.kt
@@ -10,9 +10,9 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import gq.kirmanak.mealient.AppTheme
-import gq.kirmanak.mealient.Dimens
 import gq.kirmanak.mealient.shopping_list.R
+import gq.kirmanak.mealient.ui.AppTheme
+import gq.kirmanak.mealient.ui.Dimens
 
 @Composable
 fun EmptyListError(
