+++ ui/src/main/kotlin/gq/kirmanak/mealient/ui/Theme.kt
@@ -1,4 +1,4 @@
-package gq.kirmanak.mealient
+package gq.kirmanak.mealient.ui
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
